+++ packages/types/src/entities/fields/NanoId.ts
@@ -23,6 +23,9 @@ import { z } from 'zod';
 
 import { NANOID_REGEX } from '../../common/index.js';
 
+export const ID_ALPHABET =
+	'0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz' as const;
+
 export const NanoId = z.string().regex(NANOID_REGEX);
 export type NanoId = z.infer<typeof NanoId>;
 export const NanoIdSchema: SchemaObject = generateSchema(NanoId);
